share content using system share sheet
This commit fixes #2 by removing the WRITE_EXTERNAL_STORAGE permission (which isn't granted on Android 13 and up!), and may not pass app review for the GPS.

Instead, it uses a FileProvider to provide content URIs for a file written to the internal storage of the app, and provides READ permission to the receiving app. This commit was tested against Android 13, with more tests to follow.