strong SemVer-based validation for import
The app version (BuildConfig.VersionName) is interpreted using SemVer, and if a major version change occurs, this will need to be migrated explicitly in future versions - if more stuff is serialized in the JSON than in previous versions, this will need to be taken into account. But since we are currently still on V1.x.y, no migrations are actually implemented so far.